requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Google Cloud & AI
google-cloud-aiplatform==1.75.0
google-auth==2.36.0

# Data Validation & Utilities
pydantic==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.12